Added Dynamic client Signed-off-by:Phillip Kruger <phillip.kruger@gmail.com>

+++ quarkus-client/src/main/java/com/github/phillipkruger/user/client/PersonGraphQLClient.java
@@ -1,12 +1,13 @@
 package com.github.phillipkruger.user.client;
 
 import io.smallrye.graphql.client.typesafe.api.GraphQLClientApi;
+import java.math.BigInteger;
 import org.eclipse.microprofile.graphql.Query;
 
 @GraphQLClientApi
 public interface PersonGraphQLClient {
     
     @Query
-    public Person getPerson(int id);
+    public Person getPerson(BigInteger id);
     
 }
